Move Simplifier into simplify_expressions.rs
